add `kartothek.io.eager.update_dataset_from_dataframes` function and tests
